+++ src/utils/path.rs
@@ -1,8 +1,8 @@
 use std::path::Path;
 
-pub(crate) fn path_to_root(path: &Path) -> String {
+pub(crate) fn path_to_root<P: AsRef<Path>>(path: P) -> String {
 	let mut parts = Vec::new();
-	let mut current = path.parent().unwrap();
+	let mut current = path.as_ref().parent().unwrap();
 
 	while let Some(parent) = current.parent() {
 		if current == parent {
@@ -17,5 +17,17 @@ pub(crate) fn path_to_root(path: &Path) -> String {
 		".".into()
 	} else {
 		parts.join("/")
+	}
+}
+
+#[cfg(test)]
+mod test {
+	use super::path_to_root;
+
+	#[test]
+	fn test_path_to_root() {
+		assert_eq!("..", path_to_root("lorem/ipsum.html"));
+		assert_eq!("../..", path_to_root("lorem/ipsum/dolor.html"));
+		assert_eq!(".", path_to_root("lorem.html"));
 	}
 }